kaggle_store_item_demand_forecasting/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
appnope==0.1.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
backcall==0.2.0
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cliff==4.0.0
cloudpickle==2.2.0
cmaes==0.8.2
cmd2==2.4.2
colorlog==6.7.0
commonmark==0.9.1
coverage==6.5.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
docker==6.0.0
docstring-parser==0.15
entrypoints==0.4
executing==1.1.1
fire==0.4.0
frozenlist==1.3.1
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-core==2.10.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-aiplatform==1.18.1
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-notebooks==1.4.4
google-cloud-pipeline-components==1.0.25
google-cloud-resource-manager==1.6.3
google-cloud-storage==1.44.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.48.2
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter_client==7.4.2
kaggle==1.5.12
kfp==1.8.14
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kubernetes==18.20.0
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
nose==1.3.7
numpy==1.23.4
oauthlib==3.2.1
optuna==2.10.1
packaging==21.3
pandas==1.5.0
parso==0.8.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
prettytable==3.4.1
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.4
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.1
stevedore==4.0.1
strip-hints==0.1.10
tabulate==0.9.0
termcolor==2.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typer==0.6.1
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.9.0
